refactor: simplify is_c_symbol Windows path + sync Windows-support docs

+++ examples/024_template_function/rust_hicc/src/main.rs
@@ -11,8 +11,8 @@ hicc::import_lib! {
     #[cpp(func = "void swap_double(double*, double*)")]
     unsafe fn swap_double(a: *mut f64, b: *mut f64);
 
-    #[cpp(func = "void swap_char(char*, char*)")]
-    unsafe fn swap_char(a: *mut i8, b: *mut i8);
+    #[cpp(func = "void swap_char(unsigned char*, unsigned char*)")]
+    unsafe fn swap_char(a: *mut u8, b: *mut u8);
 
     #[cpp(func = "void swap_int_array(int*, int, int)")]
     unsafe fn swap_int_array(arr: *mut i32, i: i32, j: i32);
@@ -46,11 +46,11 @@ fn main() {
     println!();
 
     // swap char
-    let mut c1 = b'A' as i8;
-    let mut c2 = b'B' as i8;
-    println!("Before swap: c1 = {}, c2 = {}", c1 as u8 as char, c2 as u8 as char);
+    let mut c1 = b'A';
+    let mut c2 = b'B';
+    println!("Before swap: c1 = {}, c2 = {}", c1 as char, c2 as char);
     unsafe { swap_char(&mut c1, &mut c2); }
-    println!("After swap: c1 = {}, c2 = {}", c1 as u8 as char, c2 as u8 as char);
+    println!("After swap: c1 = {}, c2 = {}", c1 as char, c2 as char);
 
     println!();
